refactor: change name of lang-emiter to emiters
