+++ src/ts/controllers/game-cycle/game-rounds.ts
@@ -15,7 +15,7 @@ import {
   LangEmitter,
   TempoEmitter,
   VolumeEmitter,
-} from '../emitters/lang-emitter';
+} from '../emitters/emitters';
 
 export type GameQuizViewConstructor<QuizType extends IRound = IRound> = new (
   parentNode: HTMLElement | null,
