+++ src/ts/views/components/header.ts
@@ -1,5 +1,5 @@
 import Translation from '../../constants/translation';
-import { LangEmitter } from '../../controllers/emitters/lang-emitter';
+import { LangEmitter } from '../../controllers/emitters/emitters';
 import NodeBuilder from '../../helpers/node-builder';
 import { Languages } from '../../types/data-types';
