+++ src/ts/views/components/volume-setting.ts
@@ -1,5 +1,5 @@
 import Translation from '../../constants/translation';
-import { SettingsEmitter, LangEmitter, VolumeEmitter } from '../../controllers/emitters/lang-emitter';
+import { SettingsEmitter, LangEmitter, VolumeEmitter } from '../../controllers/emitters/emitters';
 import ButtonBuilder from '../../helpers/button-builder';
 import NodeBuilder from '../../helpers/node-builder';
 import UserConfig from '../../models/user-config';
