+++ src/ts/views/pages/ear-training-page.ts
@@ -1,5 +1,5 @@
 import Translation from '../../constants/translation';
-import { LangEmitter } from '../../controllers/emitters/lang-emitter';
+import { LangEmitter } from '../../controllers/emitters/emitters';
 import GameRoundsController from '../../controllers/game-cycle/game-rounds';
 import ButtonBuilder from '../../helpers/button-builder';
 import NodeBuilder from '../../helpers/node-builder';
@@ -55,14 +55,14 @@ class EarTrainingView extends NodeBuilder {
           className: 'ear-training__btn ear-training__theory training-btn theory-btn',
           content: Translation.mainMenuTheory[state],
         }],
-        '/theory/intervals',
+          '/theory/intervals',
         ],
         [[{
           parentNode: null,
           className: 'ear-training__btn ear-training__game training-btn',
           content: Translation.intervalComparisonBtn[state],
         }],
-        '/ear-training/interval-comparison'],
+          '/ear-training/interval-comparison'],
       ],
     );
 
@@ -78,14 +78,14 @@ class EarTrainingView extends NodeBuilder {
           className: 'ear-training__btn ear-training__theory training-btn theory-btn',
           content: Translation.mainMenuTheory[state],
         }],
-        '/theory/scales',
+          '/theory/scales',
         ],
         [[{
           parentNode: null,
           className: 'ear-training__btn ear-training__game training-btn',
           content: Translation.scaleIdentificationBtn[state],
         }],
-        '/ear-training/scale-identification',
+          '/ear-training/scale-identification',
         ],
       ],
     );
@@ -102,14 +102,14 @@ class EarTrainingView extends NodeBuilder {
           className: 'ear-training__btn ear-training__theory training-btn theory-btn',
           content: Translation.mainMenuTheory[state],
         }],
-        '/theory/chords',
+          '/theory/chords',
         ],
         [[{
           parentNode: null,
           className: 'ear-training__btn ear-training__game training-btn',
           content: Translation.chordIdentificationBtn[state],
         }],
-        '/ear-training/chord-identification',
+          '/ear-training/chord-identification',
         ],
       ],
     );
